+++ tests/test_basic_s3.py
@@ -2,7 +2,7 @@ import asyncio
 import base64
 import hashlib
 from collections import defaultdict
-from typing import Callable, Literal
+from typing import Callable
 
 import aioitertools
 import botocore.retries.adaptive
@@ -24,8 +24,6 @@ async def fetch_all(pages):
     return responses
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_can_make_request(s3_client):
     # Basic smoke test to ensure we can talk to s3.
     result = await s3_client.list_buckets()
@@ -35,9 +33,7 @@ async def test_can_make_request(s3_client):
     assert actual_keys == ['Buckets', 'Owner', 'ResponseMetadata']
 
 
-@pytest.mark.moto
 @pytest.mark.parametrize('s3_verify', [False])
-@pytest.mark.asyncio
 async def test_can_make_request_no_verify(s3_client):
     # Basic smoke test to ensure we can talk to s3.
     result = await s3_client.list_buckets()
@@ -53,24 +49,22 @@ def skip_httpx(current_http_backend: str) -> None:
         pytest.skip('proxy support not implemented for httpx')
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
-async def test_fail_proxy_request(skip_httpx, aa_fail_proxy_config, s3_client):
+async def test_fail_proxy_request(
+    skip_httpx, aa_fail_proxy_config, s3_client, monkeypatch
+):
     # based on test_can_make_request
     with pytest.raises(httpsession.ProxyConnectionError):
         await s3_client.list_buckets()
 
 
-@pytest.mark.asyncio
 @pytest.mark.parametrize('mocking_test', [False])
+@pytest.mark.localonly
 async def test_succeed_proxy_request(aa_succeed_proxy_config, s3_client):
     result = await s3_client.list_buckets()
     actual_keys = sorted(list(result.keys()))
     assert actual_keys == ['Buckets', 'Owner', 'ResponseMetadata']
 
 
-@pytest.mark.asyncio
-@pytest.mark.moto
 async def test_can_get_bucket_location(s3_client, bucket_name):
     result = await s3_client.get_bucket_location(Bucket=bucket_name)
     assert 'LocationConstraint' in result
@@ -79,8 +73,6 @@ async def test_can_get_bucket_location(s3_client, bucket_name):
     assert result['LocationConstraint'] in [None, 'us-west-2', 'us-east-1']
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_can_delete_urlencoded_object(
     s3_client, bucket_name, create_object
 ):
@@ -101,8 +93,6 @@ async def test_can_delete_urlencoded_object(
     pytest.aio.assert_status_code(response, 204)
 
 
-@pytest.mark.asyncio
-@pytest.mark.moto
 async def test_can_paginate(s3_client, bucket_name, create_object):
     for i in range(5):
         key_name = f'key{i}'
@@ -117,8 +107,6 @@ async def test_can_paginate(s3_client, bucket_name, create_object):
     assert key_names == ['key0', 'key1', 'key2', 'key3', 'key4']
 
 
-@pytest.mark.asyncio
-@pytest.mark.moto
 async def test_can_paginate_with_page_size(
     s3_client, bucket_name, create_object
 ):
@@ -138,8 +126,6 @@ async def test_can_paginate_with_page_size(
     assert key_names == ['key0', 'key1', 'key2', 'key3', 'key4']
 
 
-@pytest.mark.asyncio
-@pytest.mark.moto
 async def test_can_search_paginate(s3_client, bucket_name, create_object):
     keys = []
     for i in range(5):
@@ -153,8 +139,6 @@ async def test_can_search_paginate(s3_client, bucket_name, create_object):
         assert key_name in keys
 
 
-@pytest.mark.asyncio
-@pytest.mark.moto
 async def test_can_paginate_iterator(s3_client, bucket_name, create_object):
     for i in range(5):
         key_name = f'key{i}'
@@ -173,8 +157,6 @@ async def test_can_paginate_iterator(s3_client, bucket_name, create_object):
     assert key_names == ['key0', 'key1', 'key2', 'key3', 'key4']
 
 
-@pytest.mark.asyncio
-@pytest.mark.moto
 async def test_result_key_iters(s3_client, bucket_name, create_object):
     for i in range(5):
         key_name = f'key/{i}/{i}'
@@ -204,15 +186,15 @@ async def test_result_key_iters(s3_client, bucket_name, create_object):
     assert 'CommonPrefixes' in response
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_can_get_and_put_object(
     s3_client: aiobotocore.client.AioBaseClient,
     create_object: Callable,
-    bucket_name: object,
+    bucket_name: str,
 ):
-    await create_object('foobarbaz', body='body contents')
-    resp = await s3_client.get_object(Bucket=bucket_name, Key='foobarbaz')
+    key_name = 'foobarbaz'
+    await create_object(key_name, body='body contents')
+
+    resp = await s3_client.get_object(Bucket=bucket_name, Key=key_name)
     if httpx and isinstance(resp['Body'], httpx.Response):
         data = await resp['Body'].aread()
         # note that calling `aclose()` is redundant, httpx will auto-close when the
@@ -220,12 +202,23 @@ async def test_can_get_and_put_object(
         await resp['Body'].aclose()
     else:
         data = await resp['Body'].read()
+        # TODO: think about better api and make behavior like in aiohttp
         resp['Body'].close()
     assert data == b'body contents'
 
+    # now test checksum'd file
+    key_name = 'foobarbaz_cs'
+    await create_object(key_name, b'abcd', ChecksumAlgorithm="sha1")
+    resp = await s3_client.get_object(
+        Bucket=bucket_name, Key=key_name, ChecksumMode="ENABLED"
+    )
+    if httpx and isinstance(resp['Body'], httpx.Response):
+        data = await resp['Body'].aread()
+    else:
+        data = await resp['Body'].read()
+    assert data == b'abcd'
 
-@pytest.mark.moto
-@pytest.mark.asyncio
+
 @pytest.mark.patch_attributes(
     [
         dict(
@@ -278,8 +271,6 @@ async def test_adaptive_retry(
     patch_attributes[3].assert_not_called()
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_get_object_stream_wrapper(
     s3_client, create_object, bucket_name
 ):
@@ -301,8 +292,6 @@ async def test_get_object_stream_wrapper(
     assert chunk2 == b'ody contents'
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_get_object_stream_context(
     s3_client, create_object, bucket_name
 ):
@@ -317,8 +306,6 @@ async def test_get_object_stream_context(
     assert data == b'body contents'
 
 
-@pytest.mark.asyncio
-@pytest.mark.moto
 async def test_paginate_max_items(
     s3_client, create_multipart_upload, bucket_name
 ):
@@ -358,8 +345,6 @@ async def test_paginate_max_items(
     assert len(full_result['Uploads']) == 1
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_paginate_within_page_boundaries(
     s3_client, create_object, bucket_name
 ):
@@ -402,8 +387,8 @@ async def test_paginate_within_page_boundaries(
     assert fourth['Contents'][-1]['Key'] == 'd'
 
 
-@pytest.mark.asyncio
 @pytest.mark.parametrize('mocking_test', [False])
+@pytest.mark.localonly
 async def test_unicode_key_put_list(s3_client, bucket_name, create_object):
     # Verify we can upload a key with a unicode char and list it as well.
     key_name = '\u2713'
@@ -421,8 +406,8 @@ async def test_unicode_key_put_list(s3_client, bucket_name, create_object):
     assert data == b'foo'
 
 
-@pytest.mark.asyncio
 @pytest.mark.parametrize('mocking_test', [False])
+@pytest.mark.localonly
 async def test_unicode_system_character(s3_client, bucket_name, create_object):
     # Verify we can use a unicode system character which would normally
     # break the xml parser
@@ -439,14 +424,7 @@ async def test_unicode_system_character(s3_client, bucket_name, create_object):
     assert parsed['Contents'][0]['Key'] == 'foo%08'
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
-async def test_non_normalized_key_paths(
-    s3_client,
-    bucket_name,
-    create_object,
-    current_http_backend: Literal['httpx', 'aiohttp'],
-):
+async def test_non_normalized_key_paths(s3_client, bucket_name, create_object):
     # The create_object method has assertEqual checks for 200 status.
     await create_object('key./././name')
     bucket = await s3_client.list_objects(Bucket=bucket_name)
@@ -456,7 +434,7 @@ async def test_non_normalized_key_paths(
 
 
 @pytest.mark.skipif(True, reason='Not supported')
-@pytest.mark.asyncio
+@pytest.mark.localonly
 async def test_reset_stream_on_redirects(region, create_bucket):
     # Create a bucket in a non classic region.
     bucket_name = await create_bucket(region)
@@ -464,8 +442,6 @@ async def test_reset_stream_on_redirects(region, create_bucket):
     assert bucket_name
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_copy_with_quoted_char(s3_client, create_object, bucket_name):
     key_name = 'a+b/foo'
     await create_object(key_name=key_name)
@@ -487,8 +463,6 @@ async def test_copy_with_quoted_char(s3_client, create_object, bucket_name):
     assert data == b'foo'
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_copy_with_query_string(s3_client, create_object, bucket_name):
     key_name = 'a+b/foo?notVersionid=bar'
     await create_object(key_name=key_name)
@@ -511,8 +485,6 @@ async def test_copy_with_query_string(s3_client, create_object, bucket_name):
     assert data == b'foo'
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_can_copy_with_dict_form(s3_client, create_object, bucket_name):
     key_name = 'a+b/foo?versionId=abcd'
     await create_object(key_name=key_name)
@@ -535,8 +507,6 @@ async def test_can_copy_with_dict_form(s3_client, create_object, bucket_name):
     assert data == b'foo'
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_can_copy_with_dict_form_with_version(
     s3_client, create_object, bucket_name
 ):
@@ -564,8 +534,6 @@ async def test_can_copy_with_dict_form_with_version(
     assert data == b'foo'
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_copy_with_s3_metadata(s3_client, create_object, bucket_name):
     key_name = 'foo.txt'
     await create_object(key_name=key_name)
@@ -584,7 +552,7 @@ async def test_copy_with_s3_metadata(s3_client, create_object, bucket_name):
 @pytest.mark.parametrize('signature_version', ['s3'])
 # 'Content-Disposition' not supported by moto yet
 @pytest.mark.parametrize('mocking_test', [False])
-@pytest.mark.asyncio
+@pytest.mark.localonly
 async def test_presign_with_existing_query_string_values(
     s3_client, bucket_name, aio_session, create_object
 ):
@@ -616,7 +584,7 @@ async def test_presign_with_existing_query_string_values(
 @pytest.mark.parametrize('signature_version', ['s3v4'])
 # moto host will be localhost
 @pytest.mark.parametrize('mocking_test', [False])
-@pytest.mark.asyncio
+@pytest.mark.localonly
 async def test_presign_sigv4(
     s3_client, bucket_name, aio_session, create_object
 ):
@@ -645,7 +613,7 @@ async def test_presign_sigv4(
 
 @pytest.mark.parametrize('signature_version', ['s3v4'])
 @pytest.mark.parametrize('mocking_test', [False])
-@pytest.mark.asyncio
+@pytest.mark.localonly
 async def test_can_follow_signed_url_redirect(
     alternative_s3_client, create_object, bucket_name
 ):
@@ -667,7 +635,7 @@ async def test_can_follow_signed_url_redirect(
 @pytest.mark.parametrize('region', ['eu-west-1'])
 @pytest.mark.parametrize('alternative_region', ['us-west-2'])
 @pytest.mark.parametrize('mocking_test', [False])
-@pytest.mark.asyncio
+@pytest.mark.localonly
 async def test_bucket_redirect(
     s3_client, alternative_s3_client, region, create_bucket
 ):
@@ -687,8 +655,6 @@ async def test_bucket_redirect(
 
 
 @pytest.mark.parametrize('signature_version', ['s3v4'])
-@pytest.mark.asyncio
-@pytest.mark.moto
 async def test_head_object_keys(s3_client, create_object, bucket_name):
     await create_object('foobarbaz')
 
@@ -713,8 +679,6 @@ async def test_head_object_keys(s3_client, create_object, bucket_name):
 )
 @pytest.mark.parametrize('server_scheme', ['https'])
 @pytest.mark.parametrize('s3_verify', [False])
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_put_object_sha256(s3_client, bucket_name):
     data = b'test1234'
     digest = hashlib.sha256(data).digest().hex()