+++ tests/test_config.py
@@ -11,13 +11,15 @@ from aiobotocore.session import AioSession, get_session
 from tests.mock_server import AIOServer
 
 
-# NOTE: this doesn't require moto but needs to be marked to run with coverage
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_connector_args(current_http_backend: str):
     with pytest.raises(ParamValidationError):
         # wrong type
         connector_args: dict[str, object] = dict(use_dns_cache=1)
+        AioConfig(connector_args)
+
+    with pytest.raises(ParamValidationError):
+        # wrong type
+        connector_args = dict(ttl_dns_cache="1")
         AioConfig(connector_args)
 
     with pytest.raises(ParamValidationError):
@@ -68,6 +70,8 @@ async def test_connector_args(current_http_backend: str):
         AioConfig({'resolver': True}, http_session_cls=HttpxSession)
 
     # Test valid configs:
+    AioConfig({"ttl_dns_cache": None})
+    AioConfig({"ttl_dns_cache": 1})
     AioConfig({"resolver": aiohttp.resolver.DefaultResolver()})
     AioConfig({'keepalive_timeout': None})
 
@@ -80,8 +84,6 @@ async def test_connector_args(current_http_backend: str):
     assert aio_cfg.connector_args['keepalive_timeout'] == 75
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_connector_timeout():
     session = AioSession()
     config = AioConfig(
@@ -113,8 +115,6 @@ async def test_connector_timeout():
             task2.cancel()
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_connector_timeout2():
     session = AioSession()
     config = AioConfig(
@@ -135,14 +135,11 @@ async def test_connector_timeout2():
             await resp["Body"].read()
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_get_session():
     session = get_session()
     assert isinstance(session, AioSession)
 
 
-@pytest.mark.moto
 def test_merge():
     config = AioConfig()
     other_config = AioConfig()
@@ -153,8 +150,6 @@ def test_merge():
 
 
 # Check that it's possible to specify custom http_session_cls
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_config_http_session_cls():
     class SuccessExc(Exception): ...