+++ tests/test_stubber.py
@@ -1,5 +1,3 @@
-import pytest
-
 from aiobotocore.awsrequest import AioAWSResponse
 from aiobotocore.session import AioSession
 from aiobotocore.stub import AioStubber
@@ -9,8 +7,6 @@ from .mock_server import AIOServer
 # these tests don't currently care about aiohttp vs httpx
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_add_response():
     session = AioSession()
 
@@ -38,8 +34,6 @@ async def test_add_response():
         assert stubber._queue[0]['expected_params'] == expected_params
 
 
-@pytest.mark.moto
-@pytest.mark.asyncio
 async def test_add_client_error():
     session = AioSession()